viz: the big number takes movy's preset face
Tamzen 8x16 was the wrong answer to "make it bigger" and this is why: Tamzen is
a TERMINAL face. Every one of the seven cuts on disk has 1px stems and a slashed
zero, so at 9 rows in a 15-row box it was larger and still thin.

This is the face movy draws its PRESET NUMBER with -- a Nokia 13px bitmap at
cap-height 11, 2px stems, plain bowl on the zero (src/renderer/knob.ts
drawPresetValue). 11 rows of the 15, and the widest value in the fleet
(hush1:transpose "+20") is 28px of the 30 available.

PROVENANCE IS RECORDED RATHER THAN ASSUMED. movy generated the table by
rasterising an OTF it does not vendor, so the underlying typeface is identified
only as "Nokia". The bitmap is MIT via movy -- the same relationship as
font5x3.mjs, already vendored from there -- but the face it came from is not
verified. SCH-50 exists to reduce exactly this kind of exposure, so the header
says what is and is not known instead of leaving it to be rediscovered.

The only conversion is the GAP: movy's blitter adds BIG_GAP between glyphs and
ours does not, so every advance here is theirs plus one. Twelve glyphs, matching
what bigNumberText can emit.

TWO TEST ASSERTIONS WERE PROXIES THAT ONLY HELD FOR THE OLD FONT, and both
failed on a change that improved the cell:

  * "the top third of the box is empty" encoded "a 7px glyph centred in 15
    leaves 4 clear rows". The band is now derived from the font height, since
    what is being asserted is that the number does not reach where the arc does.
  * "fewer pixels than the arc" was never the property -- a 2px-stem face
    legitimately carries MORE ink than a thin ring (84 against 56). What
    separates them is EXTENT, not quantity: the number occupies one text band,
    the arc spans the box. Bounded to the widget box, because an unbounded walk
    measures the header down to the label and reports 29 rows for both.

Suite 140 pass. Geom baseline regenerated.
