--- src/shared/param_pages/font_tamzen8x16_num.mjs
@@ -1,65 +1,0 @@
-/**
- * font_tamzen8x16_num.mjs — GENERATED by scripts/bdf_to_font.py from fonts/tamzen/Tamzen8x16r.bdf
- *
- * Tamzen (a cleaned-up Tamsyn), vendored in fonts/tamzen/. Same glyph
- * encoding as font4x5.mjs, so it blits through the same code.
- *
- * Cap height 9 rows, advance 8 px.
- */
-
-const CHARS = '0123456789+-';
-const G = [
-    [8, 0, 7, 9, 60, 66, 98, 82, 74, 70, 66, 66, 60],  /* 0 */
-    [8, 0, 7, 9, 16, 24, 20, 16, 16, 16, 16, 16, 124],  /* 1 */
-    [8, 0, 7, 9, 60, 66, 64, 64, 32, 16, 8, 4, 126],  /* 2 */
-    [8, 0, 7, 9, 126, 32, 16, 56, 64, 64, 64, 66, 60],  /* 3 */
-    [8, 0, 7, 9, 32, 48, 40, 36, 34, 126, 32, 32, 32],  /* 4 */
-    [8, 0, 7, 9, 126, 2, 2, 62, 64, 64, 64, 66, 60],  /* 5 */
-    [8, 0, 7, 9, 56, 4, 2, 2, 62, 66, 66, 66, 60],  /* 6 */
-    [8, 0, 7, 9, 126, 64, 32, 32, 16, 16, 8, 8, 8],  /* 7 */
-    [8, 0, 7, 9, 60, 66, 66, 66, 60, 66, 66, 66, 60],  /* 8 */
-    [8, 0, 7, 9, 60, 66, 66, 66, 124, 64, 64, 32, 28],  /* 9 */
-    [8, 0, 8, 9, 0, 16, 16, 16, 254, 16, 16, 16, 0],  /* + */
-    [8, 0, 7, 9, 0, 0, 0, 0, 126, 0, 0, 0, 0],  /* - */
-];
-const FALLBACK_ADV = 8;
-
-function glyphFor(ch) { const i = CHARS.indexOf(ch); return i >= 0 ? G[i] : null; }
-
-export const HEIGHT = 9;
-
-export function fontWidth(str) {
-    let w = 0; const s = String(str == null ? "" : str);
-    for (let i = 0; i < s.length; i++) { const g = glyphFor(s[i]); w += g ? g[0] : FALLBACK_ADV; }
-    return w > 0 ? w - 1 : 0;
-}
-
-export function fontPrint(ctx, x, y, str, color) {
-    let cx = x; const s = String(str == null ? "" : str);
-    for (let i = 0; i < s.length; i++) {
-        const g = glyphFor(s[i]);
-        if (!g) { cx += FALLBACK_ADV; continue; }
-        const yOff = g[1], w = g[2], h = g[3];
-        for (let row = 0; row < h; row++) {
-            const bits = g[4 + row];
-            if (!bits) continue;
-            let col = 0;
-            while (col < w) {
-                if (bits & (1 << col)) {
-                    const start = col;
-                    while (col < w && (bits & (1 << col))) col++;
-                    ctx.fillRect(cx + start, y + yOff + row, col - start, 1, color);
-                } else col++;
-            }
-        }
-        cx += g[0];
-    }
-}
-
-export function missingGlyphs(str) {
-    const out = new Set();
-    for (const ch of String(str == null ? "" : str)) if (glyphFor(ch) === null) out.add(ch);
-    return out;
-}
-
-export const MEASURE = { textWidth: fontWidth };
